+++ Assets/Scripts/IInstaller/UIInstaller.cs
@@ -21,6 +21,7 @@ namespace ElementalBlacksmithStory.Core
             builder.RegisterComponentInHierarchy<NPCDialogueView>();
             builder.RegisterComponentInHierarchy<ShopCartView>();
             builder.RegisterComponentInHierarchy<SelectShopAmountView>();
+            builder.RegisterComponentInHierarchy<ShopItemGridView>();
 
             // Presenter 등록
             builder.RegisterEntryPoint<MoneyPresenter>();
@@ -34,6 +35,7 @@ namespace ElementalBlacksmithStory.Core
             builder.RegisterEntryPoint<SettingsPresenter>().AsSelf();
             builder.RegisterEntryPoint<NPCDialoguePresenter>().AsSelf();
             builder.RegisterEntryPoint<ShopCartPresenter>().AsSelf();
+            builder.RegisterEntryPoint<ShopItemGridPresenter>().AsSelf();
         }
     }
 }